successfuly queried the data with entity. Created a new Models folder in addition to model folder to clean up the name space issue
