+++ Zoolandia/App.config
@@ -6,8 +6,8 @@
   </configSections>
 
   <connectionStrings>
-    <add name="ZoolandiaDbContext" connectionString="Data Source=(LocalDB)\\MSSQLLocalDB;AttachDbFilename='C:\Users\William\Documents\Visual Studio 2015\Projects\Zoolandia\Zoolandia\ZooTables.mdf';Integrated Security=True"  providerName="System.Data.SqlClient" />
-  </connectionStrings>                         <!--  Data Source=(localdb)\\MSSQLLocalDB;Initial Catalog=Zoolandia;Integrated Security=True;Connect Timeout=30;Encrypt=False;TrustServerCertificate=True;ApplicationIntent=ReadWrite;MultiSubnetFailover=False"  -->
+    <add name="ZoolandiaDbContext" connectionString="Data Source=(LocalDB)\MSSQLLocalDB;AttachDbFilename='C:\Users\William\Documents\Visual Studio 2015\Projects\Zoolandia\Zoolandia\ZooTables.mdf';Integrated Security=True"  providerName="System.Data.SqlClient" />
+  </connectionStrings>
 
   <startup>
     <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.5.2" />
