+++ Zoolandia/Model/HabitatFolder/Decoration.cs
