+++ Zoolandia/Model/HabitatFolder/HabitatType.cs
@@ -3,12 +3,12 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using Zoolandia.Model.Animals;
-using Zoolandia.Model.Animals.Habitats;
 
-namespace Zoolandia
+namespace Zoolandia.Model
 {
-    class Aviary : Habitat, IAviary
+    public class HabitatType
     {
+        public int IdType { get; set; }
+        public string Name { get; set; }
     }
 }
