+++ Zoolandia/Model/Taxonomy/GenusFolder/Falco.cs
@@ -3,8 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using Zoolandia.Model;
 
-namespace Zoolandia.Taxonomy
+namespace Zoolandia
 {
     class Falco : Genus
     {
