+++ Zoolandia/Model/Taxonomy/GenusFolder/Galeocerdo.cs
@@ -3,8 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using Zoolandia.Model;
 
-namespace Zoolandia.Taxonomy
+namespace Zoolandia
 {
     class Galeocerdo : Genus
     {
