+++ Zoolandia/Model/Taxonomy/SpeciesFolder/Species.cs
@@ -3,9 +3,8 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using Zoolandia.Taxonomy;
 
-namespace Zoolandia.Taxonomy.Species
+namespace Zoolandia.Model
 {
     public class Species : Taxonomy
     {
