+++ Zoolandia/Model/Taxonomy/SpeciesFolder/Species_of_Falco/Mexicanus.cs
@@ -3,17 +3,18 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using Zoolandia.Model;
 
-namespace Zoolandia.Taxonomy.Species.Species_of_Falco
+namespace Zoolandia
 {
     class Mexicanus : Species
     {
         public Mexicanus()
         {
-            this.url = "https://en.wikipedia.org/wiki/Chinese_mantis";
-            // this.genus = new Tenodera("Asia");
-            this.scientificName = "Tenodera sinensis";
-            this.commonName = "Chinese Mantis";
+            //this.url = "https://en.wikipedia.org/wiki/Chinese_mantis";
+            //// this.genus = new Tenodera("Asia");
+            //this.scientificName = "Tenodera sinensis";
+            //this.commonName = "Chinese Mantis";
         }
     }
 }
