+++ Zoolandia/Model/Taxonomy/Taxonomy.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Zoolandia.Taxonomy
+namespace Zoolandia
 {
     public class Taxonomy
     {
