+++ Zoolandia/Models/Employee.cs
@@ -4,12 +4,13 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Zoolandia.Model
+namespace Zoolandia.Models
 {
-    class Maintenance
+    public class Employee
     {
-        public int idMaintenance { get; set; }
-        public int IdHabitat { get; set; }
         public int IdEmployee { get; set; }
+        public string Name { get; set; }
+        public int Salary { get; set; }
+        public int Age { get; set; }
     }
 }
