+++ Zoolandia/Zoo.cs
@@ -3,8 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using Zoolandia.Model.Animals;
-using Zoolandia.Model.Animals.Habitats;
+using Zoolandia.Model;
 
 namespace Zoolandia
 {
