+++ Zoolandia/Zoolandia.csproj
@@ -69,27 +69,32 @@
   <ItemGroup>
     <Compile Include="AnimalHandler.cs" />
     <Compile Include="Models\Animal.cs" />
+    <Compile Include="Models\Employee.cs" />
+    <Compile Include="Models\Habitat.cs" />
+    <Compile Include="Models\HabitatType.cs" />
+    <Compile Include="Models\Maintenance.cs" />
+    <Compile Include="Models\Species.cs" />
     <Compile Include="Model\Animals\Bird.cs" />
     <Compile Include="Model\Animals\Monkey.cs" />
     <Compile Include="Model\Animals\Shark.cs" />
     <Compile Include="Model\Employees\Employee.cs" />
-    <Compile Include="Model\Habitat\Habitats\Aquatic.cs" />
-    <Compile Include="Model\Habitat\Habitats\Aviary.cs" />
-    <Compile Include="Model\Habitat\Decoration.cs" />
-    <Compile Include="Model\Habitat\HabitatType.cs" />
+    <Compile Include="Model\HabitatFolder\Habitats\Aquatic.cs" />
+    <Compile Include="Model\HabitatFolder\Habitats\Aviary.cs" />
+    <Compile Include="Model\HabitatFolder\Decoration.cs" />
+    <Compile Include="Model\HabitatFolder\HabitatType.cs" />
     <Compile Include="Model\Employees\Maintenance.cs" />
     <Compile Include="Model\ZoolandiaDbContext.cs" />
     <Compile Include="Server\HttpProcessor.cs" />
     <Compile Include="Server\HttpServer.cs" />
     <Compile Include="Model\Taxonomy\GenusFolder\Genus.cs" />
     <Compile Include="Model\Animals\Animal.cs" />
-    <Compile Include="Model\Habitat\Habitat.cs" />
+    <Compile Include="Model\HabitatFolder\Habitat.cs" />
     <Compile Include="Model\Interfaces\IAnimal.cs" />
     <Compile Include="Model\Interfaces\IAquatic.cs" />
     <Compile Include="Model\Interfaces\IAviary.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Model\Habitat\Habitats\Savanna.cs" />
+    <Compile Include="Model\HabitatFolder\Habitats\Savanna.cs" />
     <Compile Include="Model\Taxonomy\SpeciesFolder\Species.cs" />
     <Compile Include="Model\Taxonomy\GenusFolder\Falco.cs" />
     <Compile Include="Model\Taxonomy\GenusFolder\Galeocerdo.cs" />
